tests: change the template's directory path when running tests
During tests, the current working directory is set to the test file's parent directory.

For tests in the handlers package, they are run from the `handlers` folder, but the templates required for the tests are at the root directory, thus, we have to set the template dir path back up one hierarchy.